+++ ChangeLog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - glTF importer inspector
   - Removed "Node Name Method" option from glTF importer inspector. It still an option at run-time, but is always `OriginalUnique` at design-time imports.
   - `Animation` setting is disabled if built-in package animation is disabled
+- For better clarity, changed type of `Sampler` properties `minFilter`, `magFilter`, `wrapS` and `wrapT` from into to enum types and added tests
 ### Fixed
 - Works again with built-in package animation disabled (thanks [@Bersaelor][Bersaelor] for #204)